Extract libcore benchmarks to a separate folder
